Did some clean up with the files
